mcp_forge/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyyaml==6.0.1
anthropic==0.25.0
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
croniter==2.0.1
jsonschema==4.21.0
networkx==3.2.1
